requirements.txt
opencv_python==4.6.0.66
Pillow
piexif
tqdm